app/requirements.txt
Django==4.2.13
django-debug-toolbar==4.4.6
gunicorn==21.2.0
psycopg2-binary==2.9.6
celery==5.4.0
django-tailwind-cli==2.13.0
redis==5.0.7